创建文件夹
Working with Folders
It is important to understand how Git thinks about folders—it doesn’t! Git is
concerned only with files. As far as it is concerned, folders are simply a place
to store those files. Because of that, there is no way to add a folder to a
project unless it includes at least one file.
Sometimes this is a problem. For example, in many software projects there
needs to be a /build folder where automatically generated files will be saved
when compiling the software. With some systems, if you don’t have such a
folder, you’ll be unable to use the project.
Creating a Folder
A common pattern that has emerged is to create an empty file called .gitkeep
in any folder that you need to create but that doesn’t really need to have any
files in it. It seems a bit strange, but it works well and it is a well-understood
convention—so if you ever need to create a folder, just create a .gitkeep file