移动文件到上一级目录   并   重命名
Renaming a Folder
You might have guessed that just as you can’t create a folder directly, you
can’t rename it directly either. If you want to move a single file from one
folder to another, you can do that by renaming it. For example, if I wanted to
move chapter_1.md from /documentation to /new_docs, I could just go to the
view page for the chapter_1.md file, click the pencil icon to edit the file, and
at the start of the filename box type ../ to go up a folder, followed by
new_docs to create or put the file into that folder instead. However, there is
no way to just rename a folder on GitHub. You’d have to rename each of the
files in the folder one at a time to move them to the new folder.